Cleanup adn restructure directory
